+++ custom_components/yale_lock_manager/const.py
@@ -2,7 +2,7 @@
 from typing import Final
 
 DOMAIN: Final = "yale_lock_manager"
-VERSION: Final = "1.0.1.4"
+VERSION: Final = "1.0.2.0"
 
 # Configuration
 CONF_LOCK_NODE_ID: Final = "lock_node_id"